+++ scheds/rust/scx_rustland/build.rs
@@ -4,8 +4,8 @@
 fn main() {
     scx_utils::BpfBuilder::new()
         .unwrap()
-        .enable_intf("src/bpf/intf.h", "bpf_intf.rs")
-        .enable_skel("src/bpf/main.bpf.c", "bpf")
+        .enable_intf("../../../rust/scx_rustland_core/src/bpf/intf.h", "bpf_intf.rs")
+        .enable_skel("../../../rust/scx_rustland_core/src/bpf/main.bpf.c", "bpf")
         .build()
         .unwrap();
 }
